green-tech-inventory-assistant/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pandas>=2.1.0
sqlalchemy>=2.0.0
